requirements.txt
google-adk
exa-py
tavily-python
firecrawl-py
python-dotenv
fastapi
uvicorn 